Bomb Explosion
Bomb Explosion has been implemented. When the bomb is dropped it raycasts to see what it has hit, if it hits a wall piece (breakable, or not) it dosen't represent the explosion any further. It also scales based on how powerful the bomb is (how many tiles it can hit).